# xiaoyang-web-http-pb-152026/api/xiaoyang/v1/character_router.pb.go
// Code generated by https://github.com/zhufuyi/sponge, DO NOT EDIT.

package v1

import (
	context "context"
	gin "github.com/gin-gonic/gin"
	errcode "github.com/zhufuyi/sponge/pkg/errcode"
	middleware "github.com/zhufuyi/sponge/pkg/gin/middleware"
	zap "go.uber.org/zap"
	strings "strings"
)

// import packages: strings. context. errcode. middleware. zap. gin.

type CharacterServiceLogicer interface {
	CreateCharacter(ctx context.Context, req *CreateCharacterRequest) (*CreateCharacterResponse, error)
	ListCharacter(ctx context.Context, req *ListCharacterRequest) (*ListCharacterResponse, error)
	DeleteCharacter(ctx context.Context, req *DeleteCharacterRequest) (*DeleteCharacterResponse, error)
	UpdateCharacter(ctx context.Context, req *UpdateCharacterRequest) (*UpdateCharacterResponse, error)
}

type CharacterServiceOption func(*characterServiceOptions)

type characterServiceOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *characterServiceOptions) apply(opts ...CharacterServiceOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithCharacterServiceHTTPResponse() CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.isFromRPC = false
	}
}

func WithCharacterServiceRPCResponse() CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.isFromRPC = true
	}
}

func WithCharacterServiceResponser(responser errcode.Responser) CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.responser = responser
	}
}

func WithCharacterServiceLogger(zapLog *zap.Logger) CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.zapLog = zapLog
	}
}

func WithCharacterServiceErrorToHTTPCode(e ...*errcode.Error) CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.httpErrors = e
	}
}

func WithCharacterServiceRPCStatusToHTTPCode(s ...*errcode.RPCStatus) CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.rpcStatus = s
	}
}

func WithCharacterServiceWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) CharacterServiceOption {
	return func(o *characterServiceOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterCharacterServiceRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic CharacterServiceLogicer,
	opts ...CharacterServiceOption) {

	o := &characterServiceOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &characterServiceRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type characterServiceRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                CharacterServiceLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *characterServiceRouter) register() {
	r.iRouter.Handle("POST", "/v1/character", r.withMiddleware("POST", "/v1/character", r.CreateCharacter_0)...)
	r.iRouter.Handle("GET", "/v1/character/list", r.withMiddleware("GET", "/v1/character/list", r.ListCharacter_0)...)
	r.iRouter.Handle("DELETE", "/v1/character/:id", r.withMiddleware("DELETE", "/v1/character/:id", r.DeleteCharacter_0)...)
	r.iRouter.Handle("PUT", "/v1/character/:id", r.withMiddleware("PUT", "/v1/character/:id", r.UpdateCharacter_0)...)

}

func (r *characterServiceRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

var _ middleware.CtxKeyString

func (r *characterServiceRouter) CreateCharacter_0(c *gin.Context) {
	req := &CreateCharacterRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.CreateCharacter(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *characterServiceRouter) ListCharacter_0(c *gin.Context) {
	req := &ListCharacterRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListCharacter(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *characterServiceRouter) DeleteCharacter_0(c *gin.Context) {
	req := &DeleteCharacterRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteCharacter(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *characterServiceRouter) UpdateCharacter_0(c *gin.Context) {
	req := &UpdateCharacterRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateCharacter(ctx, req)
	if err != nil {
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}