# xiaoyang-web-http-pb-152026/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "xiaoyang api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/character": {
      "post": {
        "summary": "创建角色",
        "description": "创建角色",
        "operationId": "CharacterService_CreateCharacter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateCharacterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateCharacterRequest"
            }
          }
        ],
        "tags": [
          "CharacterService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/character/list": {
      "get": {
        "summary": "获取用户的角色列表",
        "description": "获取用户的角色列表",
        "operationId": "CharacterService_ListCharacter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListCharacterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "qqNumber",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CharacterService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/character/{id}": {
      "delete": {
        "summary": "删除角色",
        "description": "删除角色",
        "operationId": "CharacterService_DeleteCharacter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteCharacterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "CharacterService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新角色",
        "description": "更新角色",
        "operationId": "CharacterService_UpdateCharacter",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateCharacterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CharacterServiceUpdateCharacterBody"
            }
          }
        ],
        "tags": [
          "CharacterService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "CharacterServiceUpdateCharacterBody": {
      "type": "object",
      "properties": {
        "character": {
          "$ref": "#/definitions/v1CharacterInfo"
        }
      },
      "title": "更新角色请求"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CharacterInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "qqNumber": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "server": {
          "type": "string"
        },
        "xinfa": {
          "type": "string"
        },
        "remark": {
          "type": "string"
        }
      },
      "title": "角色信息"
    },
    "v1CreateCharacterRequest": {
      "type": "object",
      "properties": {
        "character": {
          "$ref": "#/definitions/v1CharacterInfo"
        }
      },
      "title": "创建角色请求"
    },
    "v1CreateCharacterResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "创建角色响应"
    },
...
    "v1DeleteCharacterResponse": {
      "type": "object",
      "title": "删除角色响应"
    },
...
    "v1ListCharacterResponse": {
      "type": "object",
      "properties": {
        "characterList": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CharacterInfo"
          }
        }
      },
      "title": "获取用户的角色列表响应"
    },
...
    "v1UpdateCharacterResponse": {
      "type": "object",
      "title": "更新角色响应"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Input a \"Bearer your-jwt-token\" to Value",
      "name": "Authorization",
      "in": "header"
    },
    "LeagueAuth": {
      "type": "apiKey",
      "description": "Input league key to Value Input league key to Value",
      "name": "LeagueKey",
      "in": "header"
    }
  }
}